Buttons 'fly' down when paused
The buttons are now placed at a higher location above the camera. When
the game is paused, then will translate down and stop at a certain
point. Unpausing the game will cause them to return to their original
position. Note: the y and z coordinates seem to be switched when it
comes to the editor vs. the code.